% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfi.R
\name{.dup}
\alias{.dup}
\title{Duplicate an SEXP}
\usage{
.dup(x)
}
\arguments{
\item{x}{any R object}
}
\value{
the same R object
}
\description{
This is an R wrapper around the C function \code{Rf_duplicate()}. This can be useful for
selectively duplicating only some objects passed to \code{\link{.ModernFortran}}
}
