% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfi.R
\name{.ModernFortran}
\alias{.ModernFortran}
\title{Modern Fortran Interface}
\usage{
.ModernFortran(.NAME, ..., PACKAGE)
}
\arguments{
\item{.NAME}{an object of class \code{"NativeSymbol"}, \code{"NativeSymbolInfo"}, or a
character string giving the name of the subroutine}

\item{...}{Arguments passed on to the Fortran subroutine.}

\item{PACKAGE}{passed on to \code{\link[=getNativeSymbolInfo]{getNativeSymbolInfo()}}}
}
\value{
A list of similar structure to  supplied to ..., but reflecting
changes made by the Fortran code
}
\description{
Modern Fortran Interface
}
